@@ -1,6 +1,9 @@
 {
   "name": "desktop-calendar",
-  "version": "0.1.0",
+  "version": "3.0.6",
+  "author": "LETBBI & tbvjaos510 <ghghdirk12@gmail.com, tbvjaos510@naver.com>",
+  "description": "electron-vue으로 제작한 desktop calendar",
+  "license": "MIT",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -13,53 +16,67 @@
   },
   "main": "background.js",
   "dependencies": {
-    "@vuikit/icons": "^0.8.1",
-    "@vuikit/theme": "^0.8.1",
-    "auto-launch": "^5.0.5",
-    "axios": "^0.19.2",
-    "core-js": "^3.4.4",
-    "dayspan": "^1.1.0",
-    "dayspan-vuetify": "^0.4.0",
-    "electron-disable-minimize": "2.0.0",
-    "electron-json-storage": "^4.1.8",
-    "electron-updater": "^4.2.0",
-    "fullcalendar": "^3.10.1",
-    "googleapis": "47.0.0",
-    "jquery": "^3.4.1",
-    "luxon": "^1.22.0",
-    "material-design-icons-iconfont": "^5.0.1",
-    "moment": "^2.24.0",
-    "showdown": "^1.9.1",
-    "standard": "^14.3.1",
-    "uikit": "^3.3.0",
-    "vue": "^2.6.11",
-    "vue-bus": "^1.2.1",
-    "vue-color": "^2.7.0",
-    "vue-datetime": "^1.0.0-beta.11",
+    "@ckpack/vue-color": "^1.2.0",
+    "@toast-ui/editor": "^3.2.2",
+    "@vuepic/vue-datepicker": "^6.0.2",
+    "axios": "^0.27.2",
+    "core-js": "^3.8.3",
+    "electron-disable-minimize": "^2.1.0",
+    "electron-store": "^8.0.2",
+    "electron-updater": "^5.0.5",
+    "fullcalendar": "3.10.5",
+    "global": "^4.4.0",
+    "googleapis": "^137.1.0",
+    "jquery": "^3.6.0",
+    "jquery-ui": "^1.13.2",
+    "luxon": "^3.0.1",
+    "material-design-icons-iconfont": "^6.7.0",
+    "moment": "^2.29.4",
+    "showdown": "^2.1.0",
+    "standard": "^17.0.0",
+    "tiny-emitter": "^2.1.0",
+    "uikit": "^3.15.1",
+    "vue": "^3.4.27",
+    "vue-3-linkify": "^1.1.0",
     "vue-electron": "^1.0.6",
-    "vue-linkify": "^1.0.1",
-    "vue-router": "^3.1.5",
-    "vuex": "^3.1.2",
-    "vuex-persistedstate": "^2.7.0",
-    "vuikit": "^0.8.10",
-    "weekstart": "^1.0.1"
+    "vue-router": "^4.1.2",
+    "vuex": "^4.1.0",
+    "vuex-persistedstate": "^4.1.0",
+    "weekstart": "^1.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.0",
-    "@vue/cli-plugin-eslint": "^4.1.0",
-    "@vue/cli-plugin-vuex": "^4.1.0",
-    "@vue/cli-service": "^4.1.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.0.3",
-    "electron": "^7.1.11",
-    "eslint": "^6.8.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-vue": "^6.1.2",
-    "node-gyp": "^6.0.1",
-    "node-sass": "^4.12.0",
-    "prettier": "^1.19.1",
-    "sass-loader": "^8.0.0",
-    "vue-cli-plugin-electron-builder": "^1.4.4",
-    "vue-template-compiler": "^2.6.10"
-  }
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "@electron/remote":"^2.0.8",
+    "vue-cli-plugin-electron-builder": "~2.1.1",
+    "electron": "^13.0.0",
+    "electron-devtools-installer": "^3.1.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3",
+    "sass": "^1.53.0",
+    "sass-loader": "^13.0.2"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "@babel/eslint-parser"
+    },
+    "rules": {}
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
+  ]
 }
