@@ -1,5 +1,5 @@
 <template>
-  <div class="uk-text-center uk-margin-large-top">
+  <div class="uk-text-center">
     <div class="uk-width-3-5 uk-display-inline-block">
       <img src="icon.png" alt="Desktop Calendar" id="main-icon" />
       <p class="uk-text-large uk-text-bold break-keep">
@@ -9,20 +9,19 @@
         Desktop Calendar은 Google Calendar 와 동기화 되어 여러 컴퓨터,
         핸드폰과도 동기화되는 달력입니다.
       </p>
-      <vk-button
-        type="secondary"
-        size="small"
-        class="uk-margin-top"
+      <button
+        class="uk-button uk-margin-top uk-button-secondary uk-button-small"
         uk-slideshow-item="1"
-        >시작하기</vk-button
-      >
+        >
+        시작하기
+      </button>
     </div>
   </div>
 </template>
 
 <script>
 export default {
-  name: "start"
+  name: 'set-start'
 };
 </script>
 
