@@ -1,21 +1,24 @@
 <template>
-  <h1
-    class="uk-text-center uk-margin-large-top"
-    :style="{
-      'font-size': timeStyle.size,
-      'font-weight': timeStyle.weight,
-      color: timeStyle.color
-    }"
-    style="user-select:none"
-  >
-    {{ datenow }}
-  </h1>
+  <div style="padding-right: 25%;">
+    <h1
+      class="uk-text-center uk-margin-small-top"
+      :style="{
+        'font-size': timeStyle.size,
+        'font-weight': timeStyle.weight,
+        color: timeStyle.color
+      }"
+      style="user-select:none"
+    >
+      {{ datenow }}
+    </h1>
+  </div>
 </template>
 
 <script>
-import { moment } from "fullcalendar";
+import moment from 'moment'
+
 export default {
-  name: "timer",
+  name: "main-timer",
   data() {
     return {
       datenow: "wait.."
@@ -24,9 +27,9 @@ export default {
   methods: {
     time() {
       // 현재 시간 반환
-      this.datenow = moment(new Date())
-        .locale("ko")
-        .format(this.timeMoment);
+      this.datenow = moment()
+        .locale(window.defaultLocale)
+        .format(this.timeMoment)
     }
   },
   mounted() {
@@ -36,7 +39,7 @@ export default {
       this.interval = setInterval(this.time, 1000);
     }, 1000 - new Date().getMilliseconds()); // Millisecond 맞춤
   },
-  beforeDestroy() {
+  beforeUnmout() {
     clearInterval(this.interval);
   },
   computed: {
