@@ -1,13 +1,10 @@
-import Vue from "vue";
-import Vuex from "vuex";
-import createPersistedState from "vuex-persistedstate";
+import Vuex from 'vuex'
+import createPersistedState from 'vuex-persistedstate'
 
-import modules from "./modules";
-
-Vue.use(Vuex);
+import modules from './modules'
 
 export default new Vuex.Store({
-  modules,
-  strict: process.env.NODE_ENV !== "production",
-  plugins: [createPersistedState()]
-});
+    modules,
+    strict: process.env.NODE_ENV !== 'production',
+    plugins: [createPersistedState()]
+})
